[VS-1] Change frame_interval_sec from int to float64 to support sub-second intervals
The parameter was parsed with getIntParam (strconv.Atoi), which silently
dropped decimal values like 0.1, falling back to the default of 2.
Changed to float64 across all layers: API handler, models, ffmpeg processor,
DB schema, and tests.

+++ internal/database/schema.sql
@@ -14,7 +14,7 @@ CREATE TABLE IF NOT EXISTS video_tasks (
     manifest_path TEXT,
     
     segment_sec DOUBLE PRECISION NOT NULL DEFAULT 8,
-    frame_interval_sec INTEGER NOT NULL DEFAULT 2,
+    frame_interval_sec DOUBLE PRECISION NOT NULL DEFAULT 2,
     frame_format VARCHAR(10) NOT NULL DEFAULT 'jpg',
     zip_format VARCHAR(10) NOT NULL DEFAULT 'zip',
