+++ pkg/models/task.go
@@ -45,7 +45,7 @@ type VideoTask struct {
 	
 	// Processing parameters
 	SegmentSec       float64 `db:"segment_sec" json:"segment_sec"`
-	FrameIntervalSec int    `db:"frame_interval_sec" json:"frame_interval_sec"`
+	FrameIntervalSec float64 `db:"frame_interval_sec" json:"frame_interval_sec"`
 	FrameFormat     string `db:"frame_format" json:"frame_format"`
 	ZipFormat       string `db:"zip_format" json:"zip_format"`
 	
@@ -68,7 +68,7 @@ type TaskProgress struct {
 
 type TaskParams struct {
 	SegmentSec       float64 `json:"segment_sec"`
-	FrameIntervalSec int    `json:"frame_interval_sec"`
+	FrameIntervalSec float64 `json:"frame_interval_sec"`
 	FrameFormat     string `json:"frame_format"`
 	ZipFormat       string `json:"zip_format"`
 }
